backend/requirements.txt
pandas
flask


python-pptx
json

